OC-215: added two more modifiers - SEALED and NON_SEALED; extended the mask to 64 bits (long) to fit all attributes

+++ clover-core/src/main/java/com/atlassian/clover/api/registry/ModifiersInfo.java
@@ -6,7 +6,7 @@ import java.util.Collection;
 import java.util.Map;
 
 public interface ModifiersInfo {
-    int getMask();
+    long getMask();
 
     /**
      * Returns one (or more - introduced in Java 8) annotations with a given name.
